OrdersApp/serializer.py
from rest_framework import serializers
from OrdersApp.models import Order, Customer


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = '__all__'

class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        fields = '__all__'