requirements.txt
africastalking==1.2.7
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
archspec @ file:///croot/archspec_1697725767277/work
asgiref==3.7.2
attrs==23.2.0
bcrypt==4.1.2
bidict==0.22.1
blinker==1.7.0
boltons @ file:///work/ci_py311/boltons_1677685195580/work
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
chardet==4.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
conda @ file:///croot/conda_1699303130622/work
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1700150901600/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
contextlib2==21.6.0
contourpy==1.2.1
coverage==7.4.4
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
defusedxml==0.8.0rc2
Django==5.0.4
django-admin==2.0.2
django-auth0==0.0.3
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-oidc-provider==0.8.2
django-six==1.0.5
djangorestframework==3.15.1
dlib==19.24.2
dnspython==2.5.0
docopt==0.6.2
ecdsa==0.18.0
excel-base==1.0.4
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.109.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.12.2
future==1.0.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.3
idna==2.10
imgviz==1.7.5
install==1.3.5
isoweek==1.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
kiwisolver==1.4.5
libmambapy @ file:///croot/mamba-split_1698782620632/work/libmambapy
MarkupSafe==2.1.3
matplotlib==3.8.4
multidict==6.0.5
nltk==3.8.1
nodeenv==1.8.0
num2words==0.5.13
numpy==1.26.2
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
packaging @ file:///croot/packaging_1693575174725/work
passlib==1.7.4
Pillow==10.1.0
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
powerline-status==2.7
psycopg2==2.9.9
pyasn1==0.5.1
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.20.0
pydantic==2.6.0
pydantic_core==2.16.1
pygame==2.5.2
pyjwkest==1.4.2
PyJWT==2.8.0
pymongo==4.6.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.2
pyphen==0.14.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
python-engineio==4.8.2
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.11.0
python3-openid==3.2.0
PyYAML==6.0.1
QtPy==2.4.1
regex==2023.10.3
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
safetensors==0.4.2
scapy==2.5.0
schema==0.7.5
scikit-learn==1.3.2
scipy==1.11.4
screen==1.0.1
simple-websocket==1.0.0
six==1.16.0
smart-labelme @ file:///home/edcet003/Downloads/smart-labelme-master
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
starlette==0.35.1
termcolor==2.4.0
textstat==0.7.3
threadpoolctl==3.2.0
TimeConvert==3.0.13
tokenizers==0.15.1
tqdm==4.33.0
transformers==4.37.0
truststore @ file:///croot/truststore_1695244293384/work
typing_extensions==4.9.0
tzlocal==5.2
urllib3==1.26.5
uuid==1.30
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
websploit==4.0.4
Werkzeug==3.0.1
whitenoise==6.6.0
wsproto==1.2.0
xlwt==1.3.0
yarl==1.9.4
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
